// src/common/web3/typechain-types/Web3Voting.ts
export interface Web3VotingInterface extends Interface {
  getFunction(
    nameOrSignature: "candidateCount" | "candidates" | "vote" | "voters"
  ): FunctionFragment;

  getEvent(nameOrSignatureOrTopic: "VoteEvent"): EventFragment;

  encodeFunctionData(
    functionFragment: "candidateCount",
    values?: undefined
  ): string;
  encodeFunctionData(
    functionFragment: "candidates",
    values: [BigNumberish]
  ): string;
  encodeFunctionData(functionFragment: "vote", values: [BigNumberish]): string;
  encodeFunctionData(functionFragment: "voters", values: [AddressLike]): string;

  decodeFunctionResult(
    functionFragment: "candidateCount",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "candidates", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "vote", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "voters", data: BytesLike): Result;
}

export namespace VoteEventEvent {
  export type InputTuple = [_candidateId: BigNumberish];
  export type OutputTuple = [_candidateId: bigint];
  export interface OutputObject {
    _candidateId: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface Web3Voting extends BaseContract {
  connect(runner?: ContractRunner | null): Web3Voting;
  waitForDeployment(): Promise<this>;

  interface: Web3VotingInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  candidateCount: TypedContractMethod<[], [bigint], "view">;

  candidates: TypedContractMethod<
    [arg0: BigNumberish],
    [
      [bigint, string, bigint] & { id: bigint; name: string; voteCount: bigint }
    ],
    "view"
  >;

  vote: TypedContractMethod<[_candidateId: BigNumberish], [void], "nonpayable">;

  voters: TypedContractMethod<[arg0: AddressLike], [boolean], "view">;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "candidateCount"
  ): TypedContractMethod<[], [bigint], "view">;
  getFunction(
    nameOrSignature: "candidates"
  ): TypedContractMethod<
    [arg0: BigNumberish],
    [
      [bigint, string, bigint] & { id: bigint; name: string; voteCount: bigint }
    ],
    "view"
  >;
  getFunction(
    nameOrSignature: "vote"
  ): TypedContractMethod<[_candidateId: BigNumberish], [void], "nonpayable">;
  getFunction(
    nameOrSignature: "voters"
  ): TypedContractMethod<[arg0: AddressLike], [boolean], "view">;

  getEvent(
    key: "VoteEvent"
  ): TypedContractEvent<
    VoteEventEvent.InputTuple,
    VoteEventEvent.OutputTuple,
    VoteEventEvent.OutputObject
  >;

  filters: {
    "VoteEvent(uint256)": TypedContractEvent<
      VoteEventEvent.InputTuple,
      VoteEventEvent.OutputTuple,
      VoteEventEvent.OutputObject
    >;
    VoteEvent: TypedContractEvent<
      VoteEventEvent.InputTuple,
      VoteEventEvent.OutputTuple,
      VoteEventEvent.OutputObject
    >;
  };
}